+++ buildSrc/src/main/groovy/com/uber/okbuck/composer/android/AndroidResourceRuleComposer.groovy
@@ -13,7 +13,7 @@ final class AndroidResourceRuleComposer extends AndroidBuckRuleComposer {
         // no instance
     }
 
-    static Rule compose(AndroidTarget target, List<String> extraDeps = []) {
+    static Rule compose(AndroidTarget target, List<String> extraResDeps = []) {
         List<String> resDeps = []
         resDeps.addAll(external(target.main.externalDeps.findAll { String dep ->
             dep.endsWith(".aar")
@@ -24,7 +24,7 @@ final class AndroidResourceRuleComposer extends AndroidBuckRuleComposer {
         }.collect { Target targetDep ->
             resRule(targetDep as AndroidTarget)
         })
-        resDeps.addAll(extraDeps)
+        resDeps.addAll(extraResDeps)
 
         OkBuckExtension okbuck = target.rootProject.okbuck
