+++ buildSrc/src/main/groovy/com/uber/okbuck/core/model/android/AndroidAppInstrumentationTarget.groovy
@@ -8,10 +8,10 @@ import org.gradle.api.Project
 /**
  * An Android instrumentation target
  */
-class AndroidAppInstrumentationTarget extends AndroidAppTarget {
+class AndroidAppInstrumentationTarget extends AndroidInstrumentationTarget {
 
     AndroidAppInstrumentationTarget(Project project, String name) {
-        super(project, name, true)
+        super(project, name)
     }
 
     @Override
@@ -19,39 +19,5 @@ class AndroidAppInstrumentationTarget extends AndroidAppTarget {
         return ((ApplicationVariant) project.android.applicationVariants.find {
             it.name == getMainTargetName(name)
         }).testVariant
-    }
-
-    // TODO: Update to use variant once issue solved: https://youtrack.jetbrains.com/issue/KT-23411
-    @Override
-    Scope getApt() {
-        return Scope.from(project, isKapt ? "kaptAndroidTest" : baseVariant.annotationProcessorConfiguration)
-    }
-
-    @Override
-    Scope getProvided() {
-        return Scope.from(project, baseVariant.compileConfiguration)
-    }
-
-    @Override
-    Scope getMain() {
-        return Scope.from(
-                project,
-                baseVariant.runtimeConfiguration,
-                getSources(baseVariant),
-                getJavaResources(baseVariant),
-                getJavaCompilerOptions(baseVariant))
-    }
-
-    @Override
-    Scope getTest() {
-        return Scope.from(project, null)
-    }
-
-    static String getMainTargetName(String name) {
-        return name.replaceFirst(/_test$/, '')
-    }
-
-    static String getInstrumentationTargetName(String name) {
-        return "${name}_test"
     }
 }
